Move PermissionManager and related artifacts to classloading bundle to fix temporal dependency for issue #1438
