+++ dev/com.ibm.ws.classloading/resources/com/ibm/ws/classloading/internal/resources/Java2SecurityUtilMessages_de.nlsprops
@@ -1,5 +1,5 @@
 ###############################################################################
-# Copyright (c) 2010, 2014 IBM Corporation and others.
+# Copyright (c) 2010, 2018 IBM Corporation and others.
 # All rights reserved. This program and the accompanying materials
 # are made available under the terms of the Eclipse Public License v1.0
 # which accompanies this distribution, and is available at
@@ -8,7 +8,7 @@
 # Contributors:
 #     IBM Corporation - initial API and implementation
 ###############################################################################
-#CMVCPATHNAME com.ibm.ws.security.java2sec/resources/com/ibm/ws/security/java2sec/resources/Java2SecurityUtilMessages.nlsprops
+#CMVCPATHNAME com.ibm.ws.classloading/resources/com/ibm/ws/classloading/internal/resources/Java2SecurityUtilMessages.nlsprops
 #COMPONENTPREFIX CWWKS
 #COMPONENTNAMEFOR WebSphere Java 2 Security
 #ISMESSAGEFILE TRUE
