+++ dev/com.ibm.ws.classloading_test/bnd.bnd
@@ -1,5 +1,5 @@
 #*******************************************************************************
-# Copyright (c) 2017 IBM Corporation and others.
+# Copyright (c) 2017, 2018 IBM Corporation and others.
 # All rights reserved. This program and the accompanying materials
 # are made available under the terms of the Eclipse Public License v1.0
 # which accompanies this distribution, and is available at
@@ -33,11 +33,14 @@ test.project: true
 	com.ibm.ws.kernel.metatype.helper;version=latest,\
 	com.ibm.ws.kernel.security.thread;version=latest,\
 	com.ibm.ws.logging;version=latest,\
+	org.jmock:jmock-legacy;version=2.5.0, \
+	cglib:cglib-nodep;version=2.1.0.3, \
 	org.hamcrest:hamcrest-all;version=1.3, \
 	../build.sharedResources/lib/junit/old/junit.jar;version=file, \
 	com.ibm.ws.junit.extensions;version=latest, \
 	org.jmock:jmock-junit4;strategy=exact;version=2.5.1, \
 	org.jmock:jmock;strategy=exact;version=2.5.1, \
+	com.ibm.ws.org.objenesis:objenesis;version=1.0, \
 	com.ibm.ws.adaptable.module;version=latest,\
 	com.ibm.ws.artifact;version=latest,\
     com.ibm.ws.org.apache.yoko.corba.spec.1.5;version=latest,\
