+++ dev/com.ibm.ws.jca/bnd.bnd
@@ -1,5 +1,5 @@
 #*******************************************************************************
-# Copyright (c) 2017 IBM Corporation and others.
+# Copyright (c) 2017, 2018 IBM Corporation and others.
 # All rights reserved. This program and the accompanying materials
 # are made available under the terms of the Eclipse Public License v1.0
 # which accompanies this distribution, and is available at
@@ -110,7 +110,7 @@ Service-Component: \
     configuration-policy:=require;\
     immediate:=true;\
     adaptableModuleFactory=com.ibm.wsspi.adaptable.module.AdaptableModuleFactory;\
-    permissionManager=com.ibm.ws.security.java2sec.PermissionManager;\
+    permissionManager=com.ibm.ws.classloading.java2sec.PermissionManager;\
     artifactContainerFactory=com.ibm.wsspi.artifact.factory.ArtifactContainerFactory;\
     classLoadingService=com.ibm.wsspi.classloading.ClassLoadingService;\
     resourceAdapterBundleService=com.ibm.ws.jca.rar.ResourceAdapterBundleService;\
@@ -190,5 +190,4 @@ instrument.classesExcludes: com/ibm/ws/jca/internal/resources/*.class
 	com.ibm.ws.kernel.service;version=latest,\
 	com.ibm.ws.logging.core;version=latest,\
 	com.ibm.ws.org.osgi.annotation.versioning;version=latest,\
-	com.ibm.ws.kernel.boot;version=latest,\
-	com.ibm.ws.security.java2sec;version=latest
+	com.ibm.ws.kernel.boot;version=latest
