+++ dev/com.ibm.ws.jca/src/com/ibm/ws/jca/internal/ResourceAdapterService.java
@@ -38,10 +38,10 @@ import com.ibm.websphere.ras.Tr;
 import com.ibm.websphere.ras.TraceComponent;
 import com.ibm.ws.classloading.ClassProvider;
 import com.ibm.ws.classloading.LibertyClassLoader;
+import com.ibm.ws.classloading.java2sec.PermissionManager;
 import com.ibm.ws.javaee.dd.permissions.PermissionsConfig;
 import com.ibm.ws.jca.rar.ResourceAdapterBundleService;
 import com.ibm.ws.jca.utils.xml.metatype.Metatype;
-import com.ibm.ws.security.java2sec.PermissionManager;
 import com.ibm.wsspi.adaptable.module.AdaptableModuleFactory;
 import com.ibm.wsspi.adaptable.module.Container;
 import com.ibm.wsspi.adaptable.module.UnableToAdaptException;
